+++ cmn-data-svc-server/src/test/java/org/acumos/cds/test/CdsControllerTest.java
@@ -1959,12 +1959,25 @@ public class CdsControllerTest {
 
 		MLPUser cu;
 		MLPPeer pr;
+		MLPRole cr;
+		MLPRole cr2;
 		MLPSolution cs;
 		MLPCatalog ca;
 		MLPCatalog catRes;
 
 		try {
-			cu = null;
+			cr = new MLPRole();
+			cr.setName("something or the other");
+			cr = client.createRole(cr);
+			Assert.assertNotNull(cr.getRoleId());
+			logger.info("Created role: {}", cr);
+
+			cr2 = new MLPRole();
+			cr2.setName("second string");
+			cr2 = client.createRole(cr2);
+			Assert.assertNotNull(cr2.getRoleId());
+			logger.info("Created role: {}", cr2);
+
 			cu = new MLPUser();
 			cu.setEmail("testcatalog@abc.com");
 			cu.setActive(true);
@@ -1972,6 +1985,10 @@ public class CdsControllerTest {
 			cu = client.createUser(cu);
 			Assert.assertNotNull("User ID", cu.getUserId());
 			logger.info("Created user {}", cu);
+
+			client.addUserRole(cu.getUserId(), cr.getRoleId());
+			List<MLPRole> addedRoles = client.getUserRoles(cu.getUserId());
+			Assert.assertEquals(1, addedRoles.size());
 
 			pr = new MLPPeer("Peer-Name-Test-Cat", "cat.fqdn.subject.name.a.b.c", "http://peer-api", true, false,
 					"contact", "AC");
@@ -1994,6 +2011,22 @@ public class CdsControllerTest {
 					new MLPCatalog("RS", false, "restricted catalog name", "them", "http://private.acumos.org"));
 			Assert.assertNotNull("Catalog ID", catRes.getCatalogId());
 			logger.info("Created catalog {}", catRes);
+
+			List<MLPRole> catRoles = client.getCatalogRoles(catRes.getCatalogId());
+			Assert.assertTrue(catRoles.isEmpty());
+			client.addCatalogRole(catRes.getCatalogId(), cr.getRoleId());
+			catRoles = client.getCatalogRoles(catRes.getCatalogId());
+			Assert.assertFalse(catRoles.isEmpty());
+			long roleCatCount = client.getRoleCatalogsCount(cr.getRoleId());
+			Assert.assertEquals(1L, roleCatCount);
+			List<String> updRoles = new ArrayList<>();
+			updRoles.add(cr.getRoleId());
+			updRoles.add(cr2.getRoleId());
+			client.updateCatalogRoles(catRes.getCatalogId(), updRoles);
+			catRoles = client.getCatalogRoles(catRes.getCatalogId());
+			Assert.assertEquals(2, catRoles.size());
+			List<String> accessibleCatIds = client.getUserAccessCatalogIds(cu.getUserId());
+			Assert.assertEquals(1, accessibleCatIds.size());
 
 			RestPageResponse<MLPCatalog> catalogs = client.getCatalogs(new RestPageRequest(0, 2, "name"));
 			Assert.assertNotNull(catalogs);
@@ -2136,6 +2169,24 @@ public class CdsControllerTest {
 		} catch (HttpStatusCodeException ex) {
 			logger.info("drop user fave catalog failed on bad user id as expected: {}", ex.getResponseBodyAsString());
 		}
+		try {
+			client.addCatalogRole("BOGUS", cr.getRoleId());
+			throw new Exception("Unexpected success");
+		} catch (HttpStatusCodeException ex) {
+			logger.info("add catalog role failed on bad cat ID as expected: {}", ex.getResponseBodyAsString());
+		}
+		try {
+			client.addCatalogRole(ca.getCatalogId(), "bogus");
+			throw new Exception("Unexpected success");
+		} catch (HttpStatusCodeException ex) {
+			logger.info("add catalog role failed on bad role as expected: {}", ex.getResponseBodyAsString());
+		}
+		try {
+			client.updateCatalogRoles("bogus", new ArrayList<String>());
+			throw new Exception("Unexpected success");
+		} catch (HttpStatusCodeException ex) {
+			logger.info("drop catalog role failed on bad cat id as expected: {}", ex.getResponseBodyAsString());
+		}
 
 		client.dropPeerAccessCatalog(pr.getPeerId(), catRes.getCatalogId());
 		Assert.assertEquals(0, client.getPeerAccessCatalogIds(cu.getUserId()).size());
@@ -2144,11 +2195,15 @@ public class CdsControllerTest {
 		client.dropSolutionFromCatalog(cs.getSolutionId(), ca.getCatalogId());
 		Assert.assertEquals(0, client.getSolutionsInCatalogs(new String[] { ca.getCatalogId() }, new RestPageRequest())
 				.getNumberOfElements());
+		client.dropCatalogRole(catRes.getCatalogId(), cr2.getRoleId());
+		client.dropCatalogRole(catRes.getCatalogId(), cr.getRoleId());
 		client.deleteCatalog(catRes.getCatalogId());
 		client.deleteCatalog(ca.getCatalogId());
 		client.deleteSolution(cs.getSolutionId());
 		client.deletePeer(pr.getPeerId());
 		client.deleteUser(cu.getUserId());
+		client.deleteRole(cr2.getRoleId());
+		client.deleteRole(cr.getRoleId());
 	}
 
 	@Test